+++ wxchoice.mod/glue.cpp
@@ -123,7 +123,7 @@ BEGIN_EVENT_TABLE(MaxChoice, wxChoice)
 END_EVENT_TABLE()
 
 
-MaxChoice * bmx_wxchoice_create(BBObject * maxHandle, wxWindow * parent, wxWindowID id, BBArray * array, int x, int y, int w, int h, long style) {
+MaxChoice * bmx_wxchoice_create(BBObject * maxHandle, wxWindow * parent, wxWindowID id, BBArray * array, int x, int y, int w, int h, int style) {
 	return new MaxChoice(maxHandle, parent, id, bbStringArrayTowxArrayStr(array), x, y, w, h, style);
 }
