+++ wxchoicebook.mod/glue.cpp
@@ -130,7 +130,7 @@ bool MaxChoicebookXmlHandler::CanHandle(wxXmlNode *node)
 
 // *********************************************
 
-MaxChoicebook * bmx_wxchoicebook_create(BBObject * maxHandle, wxWindow * parent, wxWindowID id, int x, int y, int w, int h, long style) {
+MaxChoicebook * bmx_wxchoicebook_create(BBObject * maxHandle, wxWindow * parent, wxWindowID id, int x, int y, int w, int h, int style) {
 	return new MaxChoicebook(maxHandle, parent, id, x, y, w, h, style);
 }
