+++ wxcontrol.mod/glue.h
@@ -34,7 +34,7 @@ extern "C" {
 	void bmx_wxcontrol_setlabeltext(wxControl * control, BBString * label);
 	int bmx_wxcontrol_setlabelmarkup(wxControl * control, BBString * label);
 
-	MaxControl * bmx_wxcontrol_create(BBObject * handle, wxWindow * parent, wxWindowID id, int x, int y, int w, int h, long style);
+	MaxControl * bmx_wxcontrol_create(BBObject * handle, wxWindow * parent, wxWindowID id, int x, int y, int w, int h, int style);
 
 }
